refactor service to add bt service
