+++ src/accounts/account.controller.ts
@@ -4,18 +4,21 @@ import { IAccountAdd } from 'src/requests/AccountAdd';
 import { BRDRequest } from 'src/requests/AuthRequest';
 import { IBTCallback } from 'src/requests/BTCallback';
 import { ISyncAccountRequest } from 'src/requests/SyncBankAccountRequest';
+import { AccountCoordinator } from './account.coordinator';
 import { AccountService } from './account.service';
-import { AccountServiceBRD } from './account.serviceBRD';
+import { BrdService } from './brd.service';
+import { BtService } from './bt.service';
 import { IBankAccount } from './models/Account';
 
 @Controller('account')
 export class AccountController {
-	constructor(private readonly service: AccountService, private readonly serviceBRD: AccountServiceBRD) {}
+	constructor(private readonly service: AccountService, private readonly serviceBRD: BrdService,
+    private coordinator: AccountCoordinator, private btService: BtService) {}
 
 	@UseGuards(JwtAuthGuard)
 	@Post('add')
 	addAccount(@Request() req: IAccountAdd): Promise<string> {
-		return this.service.addAcount(req.user.userId, req.body.bank);
+		return this.coordinator.addAcount(req.user.userId, req.body.bank);
 	}
 
 	@Post('addBRD')
@@ -25,13 +28,13 @@ export class AccountController {
 
 	@Post('btcallback')
 	async btcallback(@Body() body: IBTCallback): Promise<void> {
-		await this.service.handleBTCallback(body);
+		await this.btService.handleBTCallback(body);
 	}
 
 	@UseGuards(JwtAuthGuard)
-	@Post('sync')
+	@Post('btsync')
 	async syncAccount(@Request() req: ISyncAccountRequest): Promise<void> {
-		await this.service.syncBTAccount(req.body.accountId);
+		await this.btService.syncBTAccount(req.body.accountId);
 	}
 
   @UseGuards(JwtAuthGuard)
