+++ src/requests/AccountAdd.ts
@@ -1,4 +1,4 @@
-import { EnumBanks } from 'src/accounts/models/Account';
+import { EnumBanks } from 'src/accounts/models/Oauth';
 import { AuthRequest } from './AuthRequest';
 
 export interface IAccountAdd extends AuthRequest {
